// apps/landing-page/src/_shared/types/navigation.ts
export type Lang = 'es' | 'en' | 'fr';

export interface SubItem {
  label: string;
  path: string;
  description: string;
  icon: string;
}

export interface HeaderConstants {
  bg: string;
  bgSub: string;
  border: string;
  text: string;
  textMid: string;
  textMute: string;
  accent: string;
  accentDk: string;
  accentBg: string;
  accentLine: string;
}